Let the tests run in a native MinGW64 build
The `run.sh` script seems to assume that mingw builds are all
cross-compiles as it tries to run the built test executables
via `wine`.  This patch allows cross-compiles from linux to
MinGW to use `wine`, but allows direct execution of the test
executables when building natively with MinGW64 on Windows.